feat(keeper): extract cancellation_watchdog to shared app/lib/watchdog.py
The keeper service is TypeScript/Node.js (keeper/index.ts), so the Python
cancellation_watchdog cannot be ported directly. Instead:

1. Extract watchdog to app/lib/watchdog.py (shared module)
2. Refactor app/worker.py to import from shared module
3. Document in watchdog.py that keeper requires separate liveness
   mechanism (covered by /api/health/keeper + check_keeper_freshness)

No behavior change in worker.py.
