+++ app/worker.py
@@ -47,66 +47,7 @@ logger = logging.getLogger("worker")
 _current_cycle_stats = None
 
 
-async def cancellation_watchdog():
-    """Detect tasks stuck in CANCELLING state.
-
-    When asyncio cancels a task blocked in sync code, the task transitions
-    to CANCELLING but cancellation can't propagate. This watchdog logs at
-    2min and force-exits at 5min so Railway restarts the worker.
-    """
-    cancelling_since: dict[int, float] = {}
-    while True:
-        try:
-            await asyncio.sleep(30)
-            now = time.time()
-            live_ids = set()
-
-            for t in asyncio.all_tasks():
-                tid = id(t)
-                live_ids.add(tid)
-                state = getattr(t, "_state", None)
-
-                if state == "CANCELLING":
-                    first_seen = cancelling_since.setdefault(tid, now)
-                    age = now - first_seen
-
-                    if age > 120:
-                        coro = t.get_coro()
-                        loc = "unknown"
-                        try:
-                            if coro and coro.cr_frame:
-                                loc = f"{coro.cr_code.co_filename}:{coro.cr_frame.f_lineno}"
-                        except Exception:
-                            pass
-
-                        logger.error(
-                            f"WEDGED TASK: name={t.get_name()} "
-                            f"cancelling_for={age:.0f}s coro_at={loc}"
-                        )
-
-                        if age > 300:
-                            logger.error(
-                                f"WEDGED TASK exceeded 5min ({t.get_name()}) — "
-                                f"exiting worker for restart"
-                            )
-                            for h in logging.getLogger().handlers:
-                                try:
-                                    getattr(h, "flush")()  # logging handler, not DB
-                                except Exception:
-                                    pass
-                            os._exit(1)
-                else:
-                    cancelling_since.pop(tid, None)
-
-            for tid in list(cancelling_since.keys()):
-                if tid not in live_ids:
-                    cancelling_since.pop(tid)
-
-        except asyncio.CancelledError:
-            raise
-        except Exception as e:
-            logger.error(f"cancellation_watchdog error: {e}")
-            await asyncio.sleep(30)
+from app.lib.watchdog import cancellation_watchdog
 
 
 def _record_cycle_error(
